requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==23.0.0
pydantic==2.5.2
supabase
python-dotenv
postgrest